+++ crates/stackable-versioned-macros/src/attrs/mod.rs
@@ -1,2 +1,3 @@
-pub(crate) mod container;
+pub(crate) mod common;
 pub(crate) mod field;
+pub(crate) mod variant;